Move Python 3.11 compatibility into compat module.
